injector/requirements.txt
google-cloud-storage==2.10.0
requests==2.31.0
python-dotenv==1.0.0